# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM Nekretnine API",
        "description": "API dokumentacija za CRM sistem nekretnina",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/nekretnine": {
            "get": {
                "tags": [
                    "Nekretnine"
                ],
                "summary": "Lista nekretnina (sa slikama)",
                "operationId": "ef589a74e522b349b88763f877955289",
                "responses": {
                    "200": {
                        "description": "Lista nekretnina",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Nekretnina"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Nekretnine"
                ],
                "summary": "Kreiranje nekretnine",
                "operationId": "0dbb59922ce448930cf3a1ee7348a20e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "adresa",
                                    "tip",
                                    "cena"
                                ],
                                "properties": {
                                    "adresa": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Bulevar Oslobođenja 12, Novi Sad"
                                    },
                                    "tip": {
                                        "type": "string",
                                        "maxLength": 80,
                                        "example": "stan"
                                    },
                                    "cena": {
                                        "type": "number",
                                        "example": 95000
                                    },
                                    "status": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "dostupna",
                                        "nullable": true
                                    },
                                    "atributi": {
                                        "type": "object",
                                        "example": {
                                            "kvadratura": "55",
                                            "sprat": "3",
                                            "grejanje": "CG"
                                        },
                                        "nullable": true,
                                        "additionalProperties": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nekretnina kreirana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Nekretnina"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/nekretnine/{id}": {
            "get": {
                "tags": [
                    "Nekretnine"
                ],
                "summary": "Detalji nekretnine (sa slikama)",
                "operationId": "ba14ce9fc161d79a9055106894b4ff6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nekretnina",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Nekretnina"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nekretnina nije pronađena"
                    }
                }
            },
            "put": {
                "tags": [
                    "Nekretnine"
                ],
                "summary": "Izmena nekretnine",
                "operationId": "a796b4894cfa5e0ddc2750deb32ba7db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "adresa": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Cara Dušana 8, Beograd",
                                        "nullable": true
                                    },
                                    "tip": {
                                        "type": "string",
                                        "maxLength": 80,
                                        "example": "kuca",
                                        "nullable": true
                                    },
                                    "cena": {
                                        "type": "number",
                                        "example": 125000,
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "rezervisana",
                                        "nullable": true
                                    },
                                    "atributi": {
                                        "type": "object",
                                        "example": {
                                            "kvadratura": "120",
                                            "parking": "da"
                                        },
                                        "nullable": true,
                                        "additionalProperties": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nekretnina izmenjena",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Nekretnina"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nekretnina nije pronađena"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Nekretnine"
                ],
                "summary": "Brisanje nekretnine",
                "operationId": "5deb554c8c3351ba153f9b4ce30357d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nekretnina obrisana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nekretnina obrisana"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nekretnina nije pronađena"
                    }
                }
            }
        },
        "/api/nekretnine/search": {
            "get": {
                "tags": [
                    "Nekretnine"
                ],
                "summary": "Pretraga nekretnina (filter + paginacija + sortiranje)",
                "description": "Primer: /api/nekretnine/search?q=...&page=1&per_page=10&sort_by=adresa&sort_dir=asc&status=dostupna&tip=stan",
                "operationId": "133767be05f9eb2a2fc42e8ba0b93d86",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 150
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        },
                        "example": 10
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 50
                        },
                        "example": "dostupna"
                    },
                    {
                        "name": "tip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 80
                        },
                        "example": "stan"
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "adresa",
                                "tip",
                                "cena",
                                "status",
                                "created_at"
                            ]
                        },
                        "example": "created_at"
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        },
                        "example": "desc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginirani rezultati",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Nekretnina"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Nekretnina": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "adresa": {
                        "type": "string",
                        "example": "Bulevar Oslobođenja 12, Novi Sad"
                    },
                    "tip": {
                        "type": "string",
                        "example": "stan"
                    },
                    "cena": {
                        "type": "number",
                        "example": 95000
                    },
                    "status": {
                        "type": "string",
                        "example": "dostupna",
                        "nullable": true
                    },
                    "atributi": {
                        "type": "object",
                        "example": {
                            "kvadratura": "55",
                            "sprat": "3"
                        },
                        "nullable": true,
                        "additionalProperties": {
                            "type": "string"
                        }
                    },
                    "slike": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NekretninaSlika"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "NekretninaSlika": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "nekretnina_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "putanja": {
                        "type": "string",
                        "example": "uploads/nekretnine/1/slika1.jpg"
                    },
                    "istaknuta": {
                        "type": "boolean",
                        "example": true
                    },
                    "redosled": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}